Move List.js in preparation for TypeScript (WOR-756).
