+++ src/pages/billing/NewBillingProjectWizard/AzureBillingProjectWizard/AzureBillingProjectWizard.ts
@@ -8,7 +8,7 @@ import Events from 'src/libs/events'
 import { useOnMount } from 'src/libs/react-utils'
 import { summarizeErrors, withBusyState } from 'src/libs/utils'
 import * as Utils from 'src/libs/utils'
-import { billingProjectNameValidator } from 'src/pages/billing/List'
+import { billingProjectNameValidator } from 'src/pages/billing/Billing'
 import { AzureManagedAppCoordinates } from 'src/pages/billing/models/AzureManagedAppCoordinates'
 import { BillingRole } from 'src/pages/billing/models/BillingProject'
 import { AddUsersStep } from 'src/pages/billing/NewBillingProjectWizard/AzureBillingProjectWizard/AddUsersStep'
